+++ frame/session/benchmarking/src/mock.rs
@@ -25,7 +25,7 @@ use frame_support::parameter_types;
 type AccountId = u64;
 type AccountIndex = u32;
 type BlockNumber = u64;
-type Balance = u64;
+type Balance = u128;
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
