+++ frame/support/test/tests/pallet_compatibility_instance.rs
@@ -221,32 +221,32 @@ impl frame_system::Config for Runtime {
 impl pallet::Config for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 impl pallet::Config<pallet::Instance2> for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 impl pallet::Config<pallet::Instance3> for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 impl pallet_old::Config for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 impl pallet_old::Config<pallet_old::Instance2> for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 impl pallet_old::Config<pallet_old::Instance3> for Runtime {
 	type Event = Event;
 	type SomeConst = SomeConst;
-	type Balance = u64;
+	type Balance = u128;
 }
 
 pub type Header = sp_runtime::generic::Header<u32, sp_runtime::traits::BlakeTwo256>;
