+++ frame/vesting/src/lib.rs
@@ -447,7 +447,7 @@ mod tests {
 		pub const MaxLocks: u32 = 10;
 	}
 	impl pallet_balances::Config for Test {
-		type Balance = u64;
+		type Balance = u128;
 		type DustRemoval = ();
 		type Event = Event;
 		type ExistentialDeposit = ExistentialDeposit;
